feat: add favorite-button component
